refactor: change units and price to integer

+++ server/db/models/deposit.js
@@ -3,7 +3,7 @@ const db = require('../db')
 
 const Deposit = db.define('deposit', {
   amount: {type: Sequelize.INTEGER, allowNull: false},
-  units: {type: Sequelize.DECIMAL, allowNull: false}
+  units: {type: Sequelize.INTEGER, allowNull: false}
 })
 
 module.exports = Deposit
